backend/requirements.txt
uvicorn
python-jose[cryptography]
passlib[bcrypt]
pydantic
fastapi
